refactor(users container and form): removed form from container and create standalone component for
